+++ examples/protobuf/compiler.nix
@@ -1,0 +1,8 @@
+{ base, pkgs }:
+let
+  component = import ../../languages/rust/protobuf/compiler {
+    mkClient = base.languages.rust.mkClient;
+    protobuf = pkgs.protobuf;
+  };
+in
+(component // { path = ../../languages/rust/protobuf/compiler; })