+++ languages/python/default.nix
@@ -1,6 +1,16 @@
 { base, pkgs }:
 rec {
   mkPackage = import ./package.nix pkgs base;
+  fromProtobuf = { name, version, protoSources, protoInputs, pythonVersion ? pkgs.python3 }:
+    let
+      generatedCode = pkgs.callPackage ./protobuf.nix { inherit name version protoSources protoInputs; };
+    in
+    mkUtility {
+      inherit name version pythonVersion;
+      src = generatedCode;
+      propagatedBuildInputs = (pypkgs: [ pypkgs.grpcio ] ++ builtins.map (pi: pi.python.package) protoInputs);
+      doStandardTests = false; # We don't want to run our strict tests on generated code and stubs
+    };
   mkUtility =
     attrs@{ name
     , version
@@ -38,5 +48,36 @@ rec {
     in
     base.mkComponent {
       package = packageWithWheel;
+    };
+
+  mkClient =
+    attrs@{ name
+    , version
+    , src
+    , pythonVersion
+    , checkInputs ? (pythonPkgs: [ ])
+    , buildInputs ? (pythonPkgs: [ ])
+    , nativeBuildInputs ? (pythonPkgs: [ ])
+    , preBuild ? ""
+    , doStandardTests ? true
+    , ...
+    }:
+    let
+      package = mkPackage (attrs // {
+        inherit
+          name
+          version
+          pythonVersion
+          checkInputs
+          buildInputs
+          nativeBuildInputs
+          preBuild
+          src
+          doStandardTests;
+      });
+      application = pythonVersion.pkgs.toPythonApplication package;
+    in
+    base.mkComponent {
+      package = application;
     };
 }