+++ languages/rust/protobuf/compiler/default.nix
@@ -1,5 +1,5 @@
-{ rust, protobuf }:
-rust.mkClient {
+{ mkClient, protobuf }:
+mkClient {
   name = "rust-protobuf-compiler";
   src = ./.;
   PROTOC = "${protobuf}/bin/protoc";